* fixed SpriteSDL::Blit(), stretch variant -- sorry mips :( * this is the right place for me to drop a note about how i admire mips for coding 99% correct zoom-supporting code ... completely blindly, with improperly functioning SpriteSDL::Blit(), stretch variant  - ps. 99% because x and y are 1 pixel "off" :)  - pps and i didnt test what happens if you climb on a floor * mmzoom can now safely accept decimal values * updated downloader tools to pull 8.41 instead of 8.31 * options button now throws "not yet implemented" * logout button now prints out a small status message "logging out..." - temporarily until we get a messagebox for confirming logout with y/n * options and logout buttons now have correct "minifont" font

+++ trunk/automap.cpp
@@ -190,7 +190,7 @@ void Automap::setTileColor(int x, int y, int z, uint8_t color, uint8_t speedinde
 	}
 }
 
-void Automap::renderSelf(int x, int y, int w, int h, const Position& centerPos, int zoom)
+void Automap::renderSelf(int x, int y, int w, int h, const Position& centerPos, double zoom)
 // parameters specify where on the screen it should be painted
 {
 	//FIXME. zoom is not working because Blit doesnt scale properly the image
